+++ lib/karafka/base_controller.rb
@@ -3,55 +3,8 @@
 # Karafka module namespace
 module Karafka
   # Base controller from which all Karafka controllers should inherit
-  # Similar to Rails controllers we can define after_received callbacks
-  # that will be executed
-  #
-  # Note that if after_received return false, the chain will be stopped and
-  #   the perform method won't be executed
-  #
-  # @example Create simple controller
-  #   class ExamplesController < Karafka::BaseController
-  #     def perform
-  #       # some logic here
-  #     end
-  #   end
-  #
-  # @example Create a controller with a block after_received
-  #   class ExampleController < Karafka::BaseController
-  #     after_received do
-  #       # Here we should have some checking logic
-  #       # If false is returned, won't schedule a perform action
-  #     end
-  #
-  #     def perform
-  #       # some logic here
-  #     end
-  #   end
-  #
-  # @example Create a controller with a method after_received
-  #   class ExampleController < Karafka::BaseController
-  #     after_received :after_received_method
-  #
-  #     def perform
-  #       # some logic here
-  #     end
-  #
-  #     private
-  #
-  #     def after_received_method
-  #       # Here we should have some checking logic
-  #       # If false is returned, won't schedule a perform action
-  #     end
-  #   end
   class BaseController
     extend ActiveSupport::DescendantsTracker
-    include ActiveSupport::Callbacks
-
-    # The call method is wrapped with a set of callbacks
-    # We won't run perform at the backend if any of the callbacks
-    # returns false
-    # @see http://api.rubyonrails.org/classes/ActiveSupport/Callbacks/ClassMethods.html#method-i-get_callbacks
-    define_callbacks :after_received
 
     attr_reader :params_batch
 
@@ -65,21 +18,6 @@ module Karafka
       def topic=(topic)
         @topic = topic
         Controllers::Includer.call(self)
-      end
-
-      # Creates a callback that will be executed after receiving message but before executing the
-      #   backend for processing
-      # @param method_name [Symbol, String] method name or nil if we plan to provide a block
-      # @yield A block with a code that should be executed before scheduling
-      # @example Define a block after_received callback
-      #   after_received do
-      #     # logic here
-      #   end
-      #
-      # @example Define a class name after_received callback
-      #   after_received :method_name
-      def after_received(method_name = nil, &block)
-        set_callback :after_received, :before, method_name ? method_name : block
       end
     end
 
@@ -100,9 +38,7 @@ module Karafka
     # Executes the default controller flow, runs callbacks and if not halted
     # will call process method of a proper backend
     def call
-      run_callbacks :after_received do
-        process
-      end
+      process
     end
 
     private